+++ dtaianomaly/workflow/workflow_from_config.py
@@ -436,6 +436,9 @@ def detector_entry(entry):
     elif detector_type == "LocalPolynomialApproximation":
         return anomaly_detection.LocalPolynomialApproximation(**entry_without_type)
 
+    elif detector_type == "ChronosAnomalyDetector":
+        return anomaly_detection.ChronosAnomalyDetector(**entry_without_type)
+
     else:
         raise ValueError(f"Invalid detector entry: {entry}")